add new displayValueInBadge option
The new option will display an option's value in the badge.
If no explicit title is set, the text is used as title.
The default retains current behaviour.

My use-case (a sort of issue tracker) has fairly long texts for items,
but short IDs that are meaningful to users.
Displaying the complete text in the badge uses way too much horizontal space.